// TypeScript/resources/code/ts_lib/Apollo-Latest/apollographql/orbiter/src/generated/studio.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigInt: any;
  Blob: any;
  Date: any;
  DateTime: any;
  GraphQLDocument: any;
  JSON: any;
  Long: any;
  NaiveDateTime: any;
  Object: any;
  SHA256: any;
  StringOrInt: any;
  Timestamp: any;
  Void: any;
};
// ...
export type RoverArgumentInput = {
  key: Scalars['String'];
  value?: InputMaybe<Scalars['Object']>;
};
// ...
export const RoverTrack = gql`
    mutation RoverTrack($anonymousId: ID!, $command: String!, $cwdHash: SHA256!, $os: String!, $remoteUrlHash: SHA256, $sessionId: ID!, $version: String!, $arguments: [RoverArgumentInput!]!, $ci: String) {
  trackRoverSession(
    anonymousId: $anonymousId
    command: $command
    cwdHash: $cwdHash
    os: $os
    remoteUrlHash: $remoteUrlHash
    sessionId: $sessionId
    version: $version
    arguments: $arguments
    ci: $ci
  )
}
    `;
export type RouterTrackMutationVariables = Exact<{
  sessionId: Scalars['ID'];
  version: Scalars['String'];
  os: Scalars['String'];
  usage: Array<RouterUsageInput> | RouterUsageInput;
  ci?: InputMaybe<Scalars['String']>;
}>;


export type RouterTrackMutation = { __typename?: 'Mutation', trackRouterUsage?: any | null };

export type RoverTrackMutationVariables = Exact<{
  anonymousId: Scalars['ID'];
  command: Scalars['String'];
  cwdHash: Scalars['SHA256'];
  os: Scalars['String'];
  remoteUrlHash?: InputMaybe<Scalars['SHA256']>;
  sessionId: Scalars['ID'];
  version: Scalars['String'];
  arguments: Array<RoverArgumentInput> | RoverArgumentInput;
  ci?: InputMaybe<Scalars['String']>;
}>;


export type RoverTrackMutation = { __typename?: 'Mutation', trackRoverSession?: any | null };


export const RouterTrackDocument = gql`
    mutation RouterTrack($sessionId: ID!, $version: String!, $os: String!, $usage: [RouterUsageInput!]!, $ci: String) {
  trackRouterUsage(
    sessionId: $sessionId
    version: $version
    os: $os
    usage: $usage
    ci: $ci
  )
}
    `;
export const RoverTrackDocument = gql`
    mutation RoverTrack($anonymousId: ID!, $command: String!, $cwdHash: SHA256!, $os: String!, $remoteUrlHash: SHA256, $sessionId: ID!, $version: String!, $arguments: [RoverArgumentInput!]!, $ci: String) {
  trackRoverSession(
    anonymousId: $anonymousId
    command: $command
    cwdHash: $cwdHash
    os: $os
    remoteUrlHash: $remoteUrlHash
    sessionId: $sessionId
    version: $version
    arguments: $arguments
    ci: $ci
  )
}
    `;